% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00adductQuantif-class.R
\docType{class}
\name{adductQuantif-class}
\alias{adductQuantif-class}
\alias{show,c,adductQuantif,adductQuantif-class}
\alias{c,adductQuantif-method}
\title{adductQuantif class 
The adductQuantif class contains a peak integral matrix,
peak ranges and region of integration, the isotopic distribution identified
for each integrated peak and the target table of peaks integrated.}
\format{An object of class \code{NULL} of length 0.}
\usage{
x
}
\value{
peak integral matrix,
peak ranges and region of integration, the isotopic distribution identified
for each integrated peak and the target table of peaks integrated
and their corresponding MS1 scan isotopic patterns
}
\description{
adductQuantif class 
The adductQuantif class contains a peak integral matrix,
peak ranges and region of integration, the isotopic distribution identified
for each integrated peak and the target table of peaks integrated.
}
\section{Slots}{

\describe{
\item{\code{peakQuantTable}}{a matrix containing the peak 
integration results and 
consisting of a row for each peak
identified in each sample (e.g 200 samples
and 50 targets 200 * 50 = 10,000 rows)}

\item{\code{peakIdData}}{list of peak IDs}

\item{\code{predIsoDist}}{list of predicted Iso distances}

\item{\code{targTable}}{dataframe target table}

\item{\code{file.paths}}{character path to file}

\item{\code{Parameters}}{dataframe of specified parameters}
}}

\author{
JL Hayes \email{jlhayes1982@gmail.com}
}
\keyword{datasets}
