% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotProdSpectra.R
\name{dotProdSpectra}
\alias{dotProdSpectra}
\title{dot product calculation}
\usage{
dotProdSpectra(adductSpectra = NULL, nCores = NULL, 
minDotProdSpec = 0.8, maxGroups = 10)
}
\arguments{
\item{adductSpectra}{AdductSpec object}

\item{nCores}{numeric the number of cores to use for parallel computation. 
The default is to use 1 core.}

\item{minDotProdSpec}{numeric minimum dot product score}

\item{maxGroups}{numeric maximum number of groups to include from 
the dendrogram.}
}
\value{
adductSpectra AdductSpec object
}
\description{
hierarchical clustering 
(complete method see \code{\link[stats:hclust]{hclust}}).
Dissimilarity metric based on 1-dot product spectral similarity. 
Retention time and
mass groups are therefore further subdivided based on spectral similarity. 
If outlying mass spectra have been erroneously grouped then these will 
be reclassified.
}
