% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks.R
\name{findPeaks}
\alias{findPeaks}
\title{identify peaks}
\usage{
findPeaks(x, m = 3)
}
\arguments{
\item{x}{numeric vector of intensities.}

\item{m}{number of peaks to identify}
}
\value{
string of peaks
}
\description{
identifies peaks in a vector of intensities.
}
