% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTargTable.R
\name{generateTargTable}
\alias{generateTargTable}
\title{Make a target table for adductomicsR quantificaton using
specSimPep results}
\usage{
generateTargTable(allresultsFile = NULL, csvDir = NULL)
}
\arguments{
\item{allresultsFile}{character a full path to the allResults 
file generated by specSimPepId}

\item{csvDir}{character a full path to a directory to save the 
csv file to
output is a csv file called targTable.csv which can be used 
in the adductQuant function}
}
\value{
cvs file
}
\description{
Make a target table for adductomicsR quantificaton using
specSimPep results
}
\examples{
generateTargTable(paste0(system.file("extdata",package="adductomicsR"),
'/allResults_ALVLIAFAQYLQQCPFEDHVK_example.csv'),csvDir=getwd())
}
