% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsotopicDistributionMod.R
\name{IsotopicDistributionMod}
\alias{IsotopicDistributionMod}
\title{modified function from package OrgMassSpecR}
\usage{
IsotopicDistributionMod(formula = list(), charge = 1)
}
\arguments{
\item{formula}{list of character strings representing elemental formula}

\item{charge}{numeric for charge of the element}
}
\value{
dataframe of a spectrum
}
\description{
modified function from package OrgMassSpecR
}
\examples{
IsotopicDistributionMod(formula=list("CH3CH2OH","H2O"),charge = 1)
}
