% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loessWrapperMod.R
\name{loessWrapperMod}
\alias{loessWrapperMod}
\title{wrapper script for loess modeling}
\usage{
loessWrapperMod(x, y, span.vals = seq(0.25, 1, by = 0.05),
folds = 5)
}
\arguments{
\item{x}{predictor values}

\item{y}{response values}

\item{span.vals}{values of the tuning parameter to evaluate 
using cross validation}

\item{folds}{number of 'folds' for the cross-validation procedure}
}
\value{
LOESS model
}
\description{
adapted from bisoreg package
}
\examples{
loessWrapperMod (rnorm(200), rnorm(200), span.vals = 
seq(0.25, 1, by = 0.05),folds = 5)
}
