% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nAdjPeaks.R
\name{nAdjPeaks}
\alias{nAdjPeaks}
\title{remove lower intensity adjacent peaks}
\usage{
nAdjPeaks(peaksTmp = NULL, troughsTmp = NULL, peakRangeTmp = NULL)
}
\arguments{
\item{peaksTmp}{character vector with indices of 
detected peaks from findPeaks}

\item{troughsTmp}{character vector with indices 
of detected troughs from findPeaks}

\item{peakRangeTmp}{matrix of the peak range data 
with at least 3 columns (1. mass-to-charge, 2. intensity, 3. retention time)}
}
\value{
peaksTmp but with lower intensity adjacent 
peaks between the same troughs removed
}
\description{
remove lower intensity adjacent peaks
}
