% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakIntegrate.R
\name{peakIntegrate}
\alias{peakIntegrate}
\title{integrate a peak from a peak table with peak start and peak end 
retention times}
\usage{
peakIntegrate(peakTable = NULL, peakStart = NULL, 
peakEnd = NULL, expMass = NULL, 
expRt = NULL)
}
\arguments{
\item{peakTable}{a table of at least 5 columns:
\enumerate{
\item mass-to-charge.
\item intensity
\item adjusted retention time
\item raw retention time
\item scan numbers
}}

\item{peakStart}{retention time for peak start (in seconds).}

\item{peakEnd}{retention time for peak end (in seconds).}

\item{expMass}{expected mass-to-charge of target.}

\item{expRt}{expected retention time of target (in seconds).}
}
\value{
list with peak and peak table
}
\description{
integrate a peak from a peak table with peak start and peak end 
retention times
}
