% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probPeaks.R
\name{probPeaks}
\alias{probPeaks}
\title{potentially problematic peak identification}
\usage{
probPeaks(object = NULL, nTimesMad = 3, 
metrics = c("nMadDotProdDistN", 
"nMadSkewness", "nMadKurtosis", "nMadRtGroupDev", 
"nMadPeakArea", "duplicates"))
}
\arguments{
\item{object}{an 'AdductQuantif' class object}

\item{nTimesMad}{numeric number of median absolute 
deviations to identify potential
problem peaks.}

\item{metrics}{character string column names of metrics with 
which to identify
potential problem peaks or a list with individual nTimesMad arguments 
and with
list element names corresponding to column names of metrics.}

\item{...}{further arguments to \code{\link{mad}}}
}
\value{
'AdductQuantif' class object
}
\description{
potentially problematic peak identification
}
