% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtDevModelSave.R
\name{rtDevModelSave}
\alias{rtDevModelSave}
\title{extract and save retention time deviation models from adductSpec class 
object}
\usage{
rtDevModelSave(object = NULL, outputDir = NULL)
}
\arguments{
\item{object}{an 'adductSpec' class object or full path to a 
.RData file of the 'adductSpec' object}

\item{outputDir}{character full path to a directory to save the
.RData file
(defaults to the current working directory if unsupplied).}
}
\value{
save a .RData file containing the rt deviation models 
and returns to the
workspace.
}
\description{
extract and save retention time deviation models from adductSpec class 
object
}
