% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraCreate.R
\name{spectraCreate}
\alias{spectraCreate}
\title{Deconvolute both MS2 and MS1 levels scans adductomics}
\usage{
spectraCreate(MS2file = NULL, TICfilter = 10000, DNF = 2, minInt = 
100, minPeaks = 5)
}
\arguments{
\item{MS2file}{character vector of mzXML file locations}

\item{TICfilter}{numeric minimimum total ion current of an MS/MS scan. Any
MS/MS scan below this value will be filtered out (default=0).}

\item{DNF}{dynamic noise filter minimum signal to noise threshold 
(default = 2), calculated as the ratio between the linear model predicted 
intensity value and the actual intensity.}

\item{minInt}{numeric minimum intensity value}

\item{minPeaks}{minimum number of signal peaks following dynamic 
noise filtration (default = 5).}
}
\value{
list of MS2 spectra
}
\description{
Deconvolute both MS2 and MS1 levels scans adductomics
}
\examples{
eh = ExperimentHub();
temp = query(eh, 'adductData');
spectraCreate(MS2file = paste0(hubCache(temp),
'/data42_21221_2.mzXML'),
TICfilter = 10000, DNF= 2, minInt =
100, minPeaks = 5)
}
