% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truePeakTrough.R
\name{truePeakTrough}
\alias{truePeakTrough}
\title{true peak and trough detection}
\usage{
truePeakTrough(peaksTmp = NULL, troughsTmp = NULL, peakRangeTmp = 
NULL, minRes = 50, lrRes = FALSE)
}
\arguments{
\item{peaksTmp}{character vector with indices of detected peaks 
from findPeaks}

\item{troughsTmp}{character vector with indices of detected 
troughs from findPeaks}

\item{peakRangeTmp}{matrix of the peak range data with at least 
3 columns (1. mass-to-charge, 2. intensity, 3. retention time)}

\item{minRes}{numeric minimum percentage left/right resolution}

\item{lrRes}{logical if true both the left and right troughs 
must be above the minRes else
the peak will be discounted. (default = FALSE i.e. if only the 
left or right trough is less than minRes then the peak will be retained)}
}
\value{
a named numeric of both the peaks and troughs fitting the criteria.
}
\description{
true peak and trough detection
}
