% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00adductSpec-class.R
\docType{class}
\name{adductSpec-class}
\alias{adductSpec-class}
\alias{show,c,adductSpec-class}
\alias{x}
\alias{c,adductSpec-method}
\title{adductSpec class}
\format{An object of class \code{NULL} of length 0.}
\usage{
x
}
\value{
dynamic noise filtered composite MS/MS spectra
and their corresponding MS1 scan isotopic patterns
}
\description{
The adductSpec class contains dynamic noise filtered composite MS/MS spectra
and their corresponding MS1 scan isotopic patterns. Produced by
adductSpecGen() from mzXML files.
}
\section{Slots}{

\describe{
\item{\code{adductMS2spec}}{list of adduct MS2 spectras}

\item{\code{groupMS2spec}}{list of group MS2 spectras}

\item{\code{metaData}}{dataframe of metadata from mzXML}

\item{\code{aaResSeqs}}{matrix of amino acid sequences}

\item{\code{specPepMatches}}{list of spectra peptide matches}

\item{\code{specPepCompSpec}}{list of comp spectra peptide matches}

\item{\code{sumAdductType}}{dataframe of adduct types}

\item{\code{Peptides}}{dataframe of peptides under study}

\item{\code{rtDevModels}}{list of rtDevModels}

\item{\code{targetTable}}{dataframe target table}

\item{\code{file.paths}}{character of file path}

\item{\code{Parameters}}{dataframe of parameters}
}}

\section{Methods}{

\describe{
\item{c}{\code{signature(object = "adductSpec")}: Concatenates the 
spectra information.}
}
}

\author{
JL Hayes \email{jlhayes1982@gmail.com}
}
\keyword{datasets}
